+++ IdentitySdkCore/IdentitySdkCore/Classes/ReachFiveLogin.swift
@@ -1,0 +1,43 @@
+import Foundation
+import BrightFutures
+
+extension ReachFive {
+    
+    public func logout() -> Future<(), ReachFiveError> {
+        providers
+            .map { $0.logout() }
+            .sequence()
+            .flatMap { _ in self.reachFiveApi.logout() }
+    }
+    
+    public func refreshAccessToken(authToken: AuthToken) -> Future<AuthToken, ReachFiveError> {
+        let refreshRequest = RefreshRequest(
+            clientId: sdkConfig.clientId,
+            refreshToken: authToken.refreshToken ?? "",
+            redirectUri: sdkConfig.scheme
+        )
+        return reachFiveApi
+            .refreshAccessToken(refreshRequest)
+            .flatMap({ AuthToken.fromOpenIdTokenResponseFuture($0) })
+    }
+    
+    public func loginCallback(tkn: String, scopes: [String]?) -> Future<AuthToken, ReachFiveError> {
+        let pkce = Pkce.generate()
+        let scope = (scopes ?? scope).joined(separator: " ")
+        
+        return reachFiveApi.loginCallback(loginCallback: LoginCallback(sdkConfig: sdkConfig, scope: scope, pkce: pkce, tkn: tkn))
+            .flatMap({ self.authWithCode(code: $0, pkce: pkce) })
+    }
+    
+    internal func authWithCode(code: String, pkce: Pkce) -> Future<AuthToken, ReachFiveError> {
+        let authCodeRequest = AuthCodeRequest(
+            clientId: sdkConfig.clientId,
+            code: code,
+            redirectUri: sdkConfig.scheme,
+            pkce: pkce
+        )
+        return reachFiveApi
+            .authWithCode(authCodeRequest: authCodeRequest)
+            .flatMap({ AuthToken.fromOpenIdTokenResponseFuture($0) })
+    }
+}
